Refactor DateTime member translator for plugins
